[WIP] move option tests from StringUtilsTest to OptionTest
