+++ tests/testAutopas/tests/utils/StringUtilsTest.cpp
@@ -5,37 +5,6 @@
  */
 
 #include "StringUtilsTest.h"
-#include <autopas/options/ContainerOption.h>
-#include <autopas/options/DataLayoutOption.h>
-#include <autopas/options/SelectorStrategyOption.h>
-#include <autopas/options/TraversalOption.h>
-#include <autopas/options/TuningStrategyOption.h>
-
-TEST(StringUtilsTest, parseTraversalOptionsTest) {
-  testParseMultiple<autopas::TraversalOption>(
-      autopas::TraversalOption::getAllOptions(),
-      "c01, c04, c08, c18, c04-soa, direct, slicedv01, verletc18, verlec01, verlet-sliced, "
-      "cudac01, verletlists, c01-combined, verlet-clusters, var-verlet-lists-as-build, verlet-clusters-coloring, "
-      "verlet-cluster-cells",
-      autopas::TraversalOption::parseOptions);
-}
-
-TEST(StringUtilsTest, parseContainerOptionsTest) {
-  testParseMultiple<autopas::ContainerOption>(
-      autopas::ContainerOption::getAllOptions(),
-      "directSum, linkedCells, verletLists, verletLists-cells, vclusterlists, varVerletListsAsBuild, vclustercells",
-      autopas::ContainerOption::parseOptions);
-}
-
-TEST(StringUtilsTest, parseDataLayoutOptionsTest) {
-#if defined(AUTOPAS_CUDA)
-  auto options = "cuda, soa, aos";
-#else
-  auto options = "soa, aos";
-#endif
-  testParseMultiple<autopas::DataLayoutOption>(autopas::DataLayoutOption::getAllOptions(), options,
-                                               autopas::DataLayoutOption::parseOptions);
-}
 
 TEST(StringUtilsTest, parseDoublesTest) {
   testParseMultiple<double>({1., 1.5, 2., 3., 20.}, "1.,1.5, 2,3.00,2e1", autopas::utils::StringUtils::parseDoubles);
@@ -54,35 +23,3 @@ TEST(StringUtilsTest, parseNumberSetTest) {
   }
 }
 
-TEST(StringUtilsTest, parseSelectorOptionsTest) {
-  testParseSingle<autopas::SelectorStrategyOption>(autopas::SelectorStrategyOption::getAllOptions(),
-                                                   {"absolute", "median", "mean"},
-                                                   autopas::SelectorStrategyOption::parseOptions);
-}
-
-TEST(StringUtilsTest, parseTuningStrategyOptionsTest) {
-  testParseSingle<autopas::TuningStrategyOption>(autopas::TuningStrategyOption::getAllOptions(),
-                                                 {"full-search", "bayesian-search"},
-                                                 autopas::TuningStrategyOption::parseOptions);
-}
-
-TEST(StringUtilsTest, to_stringDataLayoutTest) {
-  testToString(autopas::DataLayoutOption::getAllOptions(), {autopas::DataLayoutOption()});
-}
-
-TEST(StringUtilsTest, to_stringSelectorStrategiesTest) {
-  testToString(autopas::SelectorStrategyOption::getAllOptions(), {autopas::SelectorStrategyOption()});
-}
-
-TEST(StringUtilsTest, to_stringContainerOptionsTest) {
-  testToString(autopas::ContainerOption::getAllOptions(), {autopas::ContainerOption()});
-}
-
-TEST(StringUtilsTest, to_stringTraversalOptionsTest) {
-  // testing for bad options does not make sense anymore
-  testToString(autopas::TraversalOption::getAllOptions(), {});
-}
-
-TEST(StringUtilsTest, to_stringTuningStrategyOptionsTest) {
-  testToString(autopas::TuningStrategyOption::getAllOptions(), {autopas::TuningStrategyOption()});
-}
